Java_20240216
Jakarta EE 에서 JSTL을 사용하려면
   1. Jakarta Standard Tag Library API
      jakarta.servlet.jsp.jstl-api-3.0.0.jar (45 KB)
   2. org.glassfish.web Jakarta Standard Tag Library Implementation:
      jakarta.servlet.jsp.jstl-3.0.1.jar (3625 KB)
설치 필요

sql 코드 작성할 때 공백 조심하기
servlet이라는 이름을 가지는 class라는 것을 기억하기
HttpServlet는 추상 클래스임.

+++ src/main/java/ch07_dao/CityDao.java
@@ -117,7 +117,7 @@ public class CityDao {
 			pstmt.setString(1, city.getName());
 			pstmt.setString(2, city.getCountryCode());
 			pstmt.setString(3, city.getDistrict());
-			pstmt.setInt((int) 4., city.getPopulation());
+			pstmt.setInt(4, city.getPopulation());
 			pstmt.setInt(5, city.getId());
 
 			pstmt.executeUpdate();
